+++ core-codemods/src/test/java/io/codemodder/codemods/LogFailedLoginCodemodTest.java
@@ -1,16 +1,16 @@
 package io.codemodder.codemods;
 
 import io.codemodder.testutils.Metadata;
-import io.codemodder.testutils.llm.LLMAssistedCodemodTest;
+import io.codemodder.testutils.llm.LLMVerifyingCodemodTestMixin;
 
 @Metadata(
     codemodType = LogFailedLoginCodemod.class,
     testResourceDir = "log-failed-login",
     dependencies = {})
-public final class LogFailedLoginCodemodTest extends LLMAssistedCodemodTest {
+public final class LogFailedLoginCodemodTest implements LLMVerifyingCodemodTestMixin {
 
   @Override
-  protected String getRequirementsPrompt() {
+  public String getRequirementsPrompt() {
     return """
         - Log a message when a login attempt fails.
         - Use the same log message style as the rest of the code.
