+++ core-codemods/src/test/java/io/codemodder/codemods/SensitiveDataLoggingCodemodTest.java
@@ -1,15 +1,16 @@
 package io.codemodder.codemods;
 
 import io.codemodder.testutils.Metadata;
-import io.codemodder.testutils.llm.LLMAssistedCodemodTest;
+import io.codemodder.testutils.llm.LLMVerifyingCodemodTestMixin;
 
 @Metadata(
     codemodType = SensitiveDataLoggingCodemod.class,
     testResourceDir = "sensitive-data-logging",
     dependencies = {})
-final class SensitiveDataLoggingCodemodTest extends LLMAssistedCodemodTest {
+final class SensitiveDataLoggingCodemodTest implements LLMVerifyingCodemodTestMixin {
+
   @Override
-  protected String getRequirementsPrompt() {
+  public String getRequirementsPrompt() {
     return """
         - Remove ALL lines that log sensitive data at INFO or higher severities.
         """;
